% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compendium_loader.R
\name{compendium_loader}
\alias{compendium_loader}
\title{Load Data from a Compendium File}
\usage{
compendium_loader(path, header = TRUE, row.names)
}
\arguments{
\item{path}{A string specifying the path to the file to be loaded.}

\item{header}{A logical value indicating whether the first row of the file
contains column headers. Defaults to \code{TRUE}.}

\item{row.names}{A vector or column index specifying the row names for the
data frame.}
}
\value{
A data frame containing the data read from the specified file.
}
\description{
This function loads data from a specified file path and returns it as a data
frame. It uses \code{read.table()} to read the file, assuming a comma-separated
values (CSV) format.
}
