% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Filter Data Based on Gene IDs and Rename Rows}
\usage{
filter(to_filter, to_keep)
}
\arguments{
\item{to_filter}{A data frame or matrix to be filtered.}

\item{to_keep}{A data frame in which the \code{Gene_ID} column specifies the
IDs of the genes to retain, while \code{Gene_Name} provides the new row names
for the filtered data.}
}
\value{
A data frame or matrix that contains only the rows corresponding to
the specified Gene IDs, with the row names updated to the corresponding
Gene Names.
}
\description{
This function filters a data frame by retaining only the rows whose row
names match the specified Gene IDs from a given set. It also renames the
rows based on the corresponding Gene Names from the \code{to_keep} data frame.
}
