% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{get_regions}
\alias{get_regions}
\title{Get Region Coordinates Based on Genomic Position}
\usage{
get_regions(row, chrom_sizes, is_circular, upstream = 300)
}
\arguments{
\item{row}{A named vector containing the genomic feature information,
such as \code{seqname} (chromosome), \code{start}, \code{end}, and \code{strand}
(strand orientation).}

\item{chrom_sizes}{A named vector or list containing the chromosome sizes,
where each chromosome name is associated with its length.}

\item{is_circular}{A logical value indicating whether the chromosome is
circular.}

\item{upstream}{An integer specifying the number of base pairs upstream
to include in the region. Default is 300.}
}
\value{
A data frame with two columns: \code{start_sequence} and \code{end_sequence},
which represent the start and end coordinates of the genomic region.
The row names indicate the chromosome.
}
\description{
This function calculates the genomic region upstream a given gene,
based on its coordinates and the chromosomal boundaries. It handles both
circular and linear chromosomes, adjusting the start and end positions for
the region accordingly. The function also considers the strand orientation
to return the appropriate upstream or downstream region.
}
