% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFM_loader.R
\name{online_PFM_loader}
\alias{online_PFM_loader}
\title{online_PFM_loader}
\usage{
online_PFM_loader(tax_id)
}
\arguments{
\item{tax_id}{A numeric or character vector representing the taxonomy ID of
the species for which to retrieve motifs (e.g., 9606 for Homo sapiens).}
}
\value{
A named list of data frames, where each data frame represents the
PFM of a motif. Each PFM contains the nucleotide frequencies ("A", "C", "G", "T")
at each position. The names of the list elements correspond to the transcription
factor names concatenated with their motif IDs (e.g., "TF_name_MA0001").
}
\description{
This function queries the JASPAR database to retrieve position frequency
matrices (PFMs)
for transcription factor binding motifs of a given species, identified by
its taxonomy ID.
}
