% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_first_genes.R
\name{rename_first_column}
\alias{rename_first_column}
\title{Rename the First Column of a Data Frame}
\usage{
rename_first_column(dataframe)
}
\arguments{
\item{dataframe}{A data frame where the first column (\code{V1}) has to be
renamed.}
}
\value{
A data frame with the first column (\code{V1}) updated to include only
the part of each value after the first period (\code{.}).
}
\description{
This function renames the values in the first column of a data frame by
extracting the portion after the first period (\code{.}).
}
