% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_first_genes.R
\name{tuGenes_filter}
\alias{tuGenes_filter}
\title{Filter the \code{X5.tuGenes} Column of a Data Frame}
\usage{
tuGenes_filter(x)
}
\arguments{
\item{x}{A data frame containing a column named \code{X5.tuGenes}.}
}
\value{
A data frame with the \code{X5.tuGenes} column updated to include only
the portion before the first semicolon (\verb{;}).
}
\description{
This function processes the \code{X5.tuGenes} column of a data frame by
extracting the portion before the first semicolon (\verb{;}).
If the column does not exist, the function stops with an error.
}
